+++ src/store/actionCreator.ts
@@ -27,6 +27,24 @@ export function removeTask(task: ITask) {
   return createActionDispatcher(action)
 }
 
+export function setEditedTask(task: ITask) {
+  const action: TaskAction = {
+    type: actionTypes.SET_EDITED_TASK,
+    task,
+  }
+
+  return createActionDispatcher(action)
+}
+
+export function finishEditingTask(task: ITask) {
+  const action: TaskAction = {
+    type: actionTypes.FINISH_EDITING_TASK,
+    task,
+  }
+
+  return createActionDispatcher(action)
+}
+
 export const createActionDispatcher = (action: TaskAction) => {
   return (dispatch: DispatchType) => dispatch(action)
 }