+++ src/store/actionTypes.ts
@@ -1,3 +1,5 @@
 export const ADD_TASK = 'ADD_TASK';
 export const UPDATE_TASK = 'UPDATE_TASK';
 export const REMOVE_TASK = 'REMOVE_TASK';
+export const SET_EDITED_TASK = 'SET_EDITED_TASK';
+export const FINISH_EDITING_TASK = 'FINISH_EDITING_TASK';